Day 7, part 1: Amplification Circuit
This one was a bit more tricky and required a pure effects
interpreter.  Also fixed a major bug in the Input instruction which was
storing in the wrong place.  Improved the print out of instructions.
Working towards having a full disasembler, but as the runtime can change
the 'code', this isn't really a good option to consider.  The docode as
you run is quite nice though.