Day 15 - Oxygen System
Major change here: the opcode machine is definitely 'finished' and
so it's moved entirely to the OpCodeMachine.hs file.

I'll probably refactor the logging out of it and make it more
performant if needed.  Currently, it runs okay for solving the Day15
problem.

Speaking of which, this is the Day15 part 1 solver.  It finds the
Oxygen system and explores the whole map.  There's only one path (in my
map) to the Oxygen system, but there might have been more, and the
program allows for that.

This took quite a long time to solve!