Finally a version of Day18 part 1 that passes
It's still quite slow though, which is annoying.  However, it does
manage to solve the maze in 50 seconds on my Desktop.  I'm sure there
are better algorithms.  In particularly, my culling algorithm probably
doesn't cull enough paths early on, and it hangs onto memory too much,
so it ends up using a lot of memory to solve the maze.