Day 18, Part 2: another failed experiment
This time, although it can solve the smaller mazes, using the same
technique as Part 1, it either takes too long (using lists) or has a
space leak (using HasgPSQ) when storing the variations of the options at
each key level.  And that means it uses 12GB of RAM and then segfaults.

I'm going to have to try another way of solving part 2.